server/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==1.4.25
databases==0.5.5
python-dotenv
aiosqlite
dotenv
